pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Ramil Suleymanov <lllumineux@outlook.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "^3.2.9"
djangorestframework = "^3.12.4"
transliterate = "^1.10.2"
python-dotenv = "^0.19.2"
psycopg2-binary = "^2.9.2"
whitenoise = "^5.3.0"
gunicorn = "^20.1.0"
djangorestframework-simplejwt = "^5.0.0"
django-cors-headers = "^3.10.1"
drf-yasg = "^1.20.0"
sentry-sdk = "^1.5.10"
celery = "^5.2.6"
redis = "^4.2.2"
pytest = "^7.1.1"
pytest-cov = "^3.0.0"
click = "^8.1.2"
django-mock-queries = "^2.1.7"
pytest-django = "^4.5.2"
pytest-mock = "^3.7.0"
factory-boy = "^3.2.1"

[tool.poetry.dev-dependencies]

[tool.black]
line-length = 120
skip-string-normalization = true

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "med_communication_platform.settings"
filterwarnings = [
    "error",
    "ignore::UserWarning",
    # note the use of single quote below to denote "raw" strings in TOML
    'ignore:function ham\(\) is deprecated:DeprecationWarning',
]
python_files = ['test_*.py']
addopts = ['-vv', '--cov', '-p', 'no:warnings']

[tool.coverage.run]
source = ["apps"]
omit = ["*/migrations/*"]